+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonLinkWrapLabel.cs
@@ -527,7 +527,7 @@ namespace Krypton.Toolkit
             // Layout and repaint with new settings
             NeedPaint(true);
 
-            PaletteChanged.Invoke(this, e);
+            PaletteChanged?.Invoke(this, e);
         }
 
         /// <summary>