+++ Source/Krypton Components/TestForm/Form1.Designer.cs
@@ -28,342 +28,321 @@
         /// </summary>
         private void InitializeComponent()
         {
-            components = new System.ComponentModel.Container();
-            kryptonPanel1 = new Krypton.Toolkit.KryptonPanel();
-            kryptonButton4 = new Krypton.Toolkit.KryptonButton();
-            kbtnVisualStudio2010Theme = new Krypton.Toolkit.KryptonButton();
-            kchkUseProgressValueAsText = new Krypton.Toolkit.KryptonCheckBox();
-            kryptonProgressBar1 = new Krypton.Toolkit.KryptonProgressBar();
-            ktrkProgressValues = new Krypton.Toolkit.KryptonTrackBar();
-            kryptonButton3 = new Krypton.Toolkit.KryptonButton();
-            kbtnIntegratedToolbar = new Krypton.Toolkit.KryptonButton();
-            kbtnTestMessagebox = new Krypton.Toolkit.KryptonButton();
-            kryptonButton2 = new Krypton.Toolkit.KryptonButton();
-            kryptonThemeComboBox1 = new Krypton.Toolkit.KryptonThemeComboBox();
-            kryptonButton1 = new Krypton.Toolkit.KryptonButton();
-            kryptonContextMenu1 = new Krypton.Toolkit.KryptonContextMenu();
-            kryptonContextMenuItems1 = new Krypton.Toolkit.KryptonContextMenuItems();
-            kryptonContextMenuItem1 = new Krypton.Toolkit.KryptonContextMenuItem();
-            kryptonContextMenuItem2 = new Krypton.Toolkit.KryptonContextMenuItem();
-            kryptonContextMenuItem3 = new Krypton.Toolkit.KryptonContextMenuItem();
-            kryptonListBox1 = new Krypton.Toolkit.KryptonListBox();
-            kryptonTextBox1 = new Krypton.Toolkit.KryptonTextBox();
-            textBox1 = new System.Windows.Forms.TextBox();
-            kryptonLanguageManager1 = new Krypton.Toolkit.KryptonLanguageManager();
-            kryptonCustomPaletteBase1 = new Krypton.Toolkit.KryptonCustomPaletteBase(components);
-            kcmdMessageboxTest = new Krypton.Toolkit.KryptonCommand();
-            buttonSpecAny1 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny2 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny3 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny4 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny5 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny6 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny7 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny8 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny9 = new Krypton.Toolkit.ButtonSpecAny();
-            buttonSpecAny10 = new Krypton.Toolkit.ButtonSpecAny();
-            kryptonIntegratedToolbarPrintCommand1 = new Krypton.Toolkit.KryptonIntegratedToolbarPrintCommand();
-            kryptonManager1 = new Krypton.Toolkit.KryptonManager(components);
-            ((System.ComponentModel.ISupportInitialize)kryptonPanel1).BeginInit();
-            kryptonPanel1.SuspendLayout();
-            ((System.ComponentModel.ISupportInitialize)kryptonThemeComboBox1).BeginInit();
-            SuspendLayout();
+            this.components = new System.ComponentModel.Container();
+            this.kryptonPanel1 = new Krypton.Toolkit.KryptonPanel();
+            this.kryptonButton4 = new Krypton.Toolkit.KryptonButton();
+            this.kbtnVisualStudio2010Theme = new Krypton.Toolkit.KryptonButton();
+            this.kchkUseProgressValueAsText = new Krypton.Toolkit.KryptonCheckBox();
+            this.kryptonProgressBar1 = new Krypton.Toolkit.KryptonProgressBar();
+            this.ktrkProgressValues = new Krypton.Toolkit.KryptonTrackBar();
+            this.kryptonButton3 = new Krypton.Toolkit.KryptonButton();
+            this.kbtnIntegratedToolbar = new Krypton.Toolkit.KryptonButton();
+            this.kbtnTestMessagebox = new Krypton.Toolkit.KryptonButton();
+            this.kryptonButton2 = new Krypton.Toolkit.KryptonButton();
+            this.kryptonThemeComboBox1 = new Krypton.Toolkit.KryptonThemeComboBox();
+            this.kryptonButton1 = new Krypton.Toolkit.KryptonButton();
+            this.kryptonContextMenu1 = new Krypton.Toolkit.KryptonContextMenu();
+            this.kryptonContextMenuItems1 = new Krypton.Toolkit.KryptonContextMenuItems();
+            this.kryptonContextMenuItem1 = new Krypton.Toolkit.KryptonContextMenuItem();
+            this.kryptonContextMenuItem2 = new Krypton.Toolkit.KryptonContextMenuItem();
+            this.kryptonContextMenuItem3 = new Krypton.Toolkit.KryptonContextMenuItem();
+            this.kryptonListBox1 = new Krypton.Toolkit.KryptonListBox();
+            this.kryptonTextBox1 = new Krypton.Toolkit.KryptonTextBox();
+            this.textBox1 = new System.Windows.Forms.TextBox();
+            this.kryptonLanguageManager1 = new Krypton.Toolkit.KryptonLanguageManager();
+            this.kryptonCustomPaletteBase1 = new Krypton.Toolkit.KryptonCustomPaletteBase(this.components);
+            this.kcmdMessageboxTest = new Krypton.Toolkit.KryptonCommand();
+            this.buttonSpecAny1 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny2 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny3 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny4 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny5 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny6 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny7 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny8 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny9 = new Krypton.Toolkit.ButtonSpecAny();
+            this.buttonSpecAny10 = new Krypton.Toolkit.ButtonSpecAny();
+            this.kryptonIntegratedToolbarPrintCommand1 = new Krypton.Toolkit.KryptonIntegratedToolbarPrintCommand();
+            this.kryptonManager1 = new Krypton.Toolkit.KryptonManager(this.components);
+            this.kryptonButton5 = new Krypton.Toolkit.KryptonButton();
+            ((System.ComponentModel.ISupportInitialize)(this.kryptonPanel1)).BeginInit();
+            this.kryptonPanel1.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.kryptonThemeComboBox1)).BeginInit();
+            this.SuspendLayout();
             // 
             // kryptonPanel1
             // 
-            kryptonPanel1.Controls.Add(kryptonButton4);
-            kryptonPanel1.Controls.Add(kbtnVisualStudio2010Theme);
-            kryptonPanel1.Controls.Add(kchkUseProgressValueAsText);
-            kryptonPanel1.Controls.Add(kryptonProgressBar1);
-            kryptonPanel1.Controls.Add(ktrkProgressValues);
-            kryptonPanel1.Controls.Add(kryptonButton3);
-            kryptonPanel1.Controls.Add(kbtnIntegratedToolbar);
-            kryptonPanel1.Controls.Add(kbtnTestMessagebox);
-            kryptonPanel1.Controls.Add(kryptonButton2);
-            kryptonPanel1.Controls.Add(kryptonThemeComboBox1);
-            kryptonPanel1.Controls.Add(kryptonButton1);
-            kryptonPanel1.Controls.Add(kryptonListBox1);
-            kryptonPanel1.Controls.Add(kryptonTextBox1);
-            kryptonPanel1.Controls.Add(textBox1);
-            kryptonPanel1.Dock = System.Windows.Forms.DockStyle.Fill;
-            kryptonPanel1.Location = new System.Drawing.Point(0, 0);
-            kryptonPanel1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
-            kryptonPanel1.Name = "kryptonPanel1";
-            kryptonPanel1.Size = new System.Drawing.Size(800, 693);
-            kryptonPanel1.TabIndex = 0;
+            this.kryptonPanel1.Controls.Add(this.kryptonButton5);
+            this.kryptonPanel1.Controls.Add(this.kryptonButton4);
+            this.kryptonPanel1.Controls.Add(this.kbtnVisualStudio2010Theme);
+            this.kryptonPanel1.Controls.Add(this.kchkUseProgressValueAsText);
+            this.kryptonPanel1.Controls.Add(this.kryptonProgressBar1);
+            this.kryptonPanel1.Controls.Add(this.ktrkProgressValues);
+            this.kryptonPanel1.Controls.Add(this.kryptonButton3);
+            this.kryptonPanel1.Controls.Add(this.kbtnIntegratedToolbar);
+            this.kryptonPanel1.Controls.Add(this.kbtnTestMessagebox);
+            this.kryptonPanel1.Controls.Add(this.kryptonButton2);
+            this.kryptonPanel1.Controls.Add(this.kryptonThemeComboBox1);
+            this.kryptonPanel1.Controls.Add(this.kryptonButton1);
+            this.kryptonPanel1.Controls.Add(this.kryptonListBox1);
+            this.kryptonPanel1.Controls.Add(this.kryptonTextBox1);
+            this.kryptonPanel1.Controls.Add(this.textBox1);
+            this.kryptonPanel1.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.kryptonPanel1.Location = new System.Drawing.Point(0, 0);
+            this.kryptonPanel1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.kryptonPanel1.Name = "kryptonPanel1";
+            this.kryptonPanel1.Size = new System.Drawing.Size(600, 563);
+            this.kryptonPanel1.TabIndex = 0;
             // 
             // kryptonButton4
             // 
-            kryptonButton4.Location = new System.Drawing.Point(17, 240);
-            kryptonButton4.Margin = new System.Windows.Forms.Padding(4);
-            kryptonButton4.Name = "kryptonButton4";
-            kryptonButton4.Size = new System.Drawing.Size(247, 31);
-            kryptonButton4.TabIndex = 14;
-            kryptonButton4.Values.Text = "Form 4";
-            kryptonButton4.Click += kryptonButton4_Click;
+            this.kryptonButton4.Location = new System.Drawing.Point(13, 195);
+            this.kryptonButton4.Name = "kryptonButton4";
+            this.kryptonButton4.Size = new System.Drawing.Size(185, 25);
+            this.kryptonButton4.TabIndex = 14;
+            this.kryptonButton4.Values.Text = "Form 4";
             // 
             // kbtnVisualStudio2010Theme
             // 
-            kbtnVisualStudio2010Theme.Location = new System.Drawing.Point(19, 396);
-            kbtnVisualStudio2010Theme.Margin = new System.Windows.Forms.Padding(4);
-            kbtnVisualStudio2010Theme.Name = "kbtnVisualStudio2010Theme";
-            kbtnVisualStudio2010Theme.Size = new System.Drawing.Size(245, 31);
-            kbtnVisualStudio2010Theme.TabIndex = 13;
-            kbtnVisualStudio2010Theme.Values.Text = "Visual Studio 2010 Theme (Form5)";
-            kbtnVisualStudio2010Theme.Click += kbtnVisualStudio2010Theme_Click;
+            this.kbtnVisualStudio2010Theme.Location = new System.Drawing.Point(14, 322);
+            this.kbtnVisualStudio2010Theme.Name = "kbtnVisualStudio2010Theme";
+            this.kbtnVisualStudio2010Theme.Size = new System.Drawing.Size(184, 25);
+            this.kbtnVisualStudio2010Theme.TabIndex = 13;
+            this.kbtnVisualStudio2010Theme.Values.Text = "Visual Studio 2010 Theme (Form5)";
             // 
             // kchkUseProgressValueAsText
             // 
-            kchkUseProgressValueAsText.Location = new System.Drawing.Point(269, 528);
-            kchkUseProgressValueAsText.Margin = new System.Windows.Forms.Padding(4);
-            kchkUseProgressValueAsText.Name = "kchkUseProgressValueAsText";
-            kchkUseProgressValueAsText.Size = new System.Drawing.Size(202, 24);
-            kchkUseProgressValueAsText.TabIndex = 12;
-            kchkUseProgressValueAsText.Values.Text = "Use progress value as text";
-            kchkUseProgressValueAsText.CheckedChanged += kchkUseProgressValueAsText_CheckedChanged;
+            this.kchkUseProgressValueAsText.Location = new System.Drawing.Point(202, 429);
+            this.kchkUseProgressValueAsText.Name = "kchkUseProgressValueAsText";
+            this.kchkUseProgressValueAsText.Size = new System.Drawing.Size(165, 20);
+            this.kchkUseProgressValueAsText.TabIndex = 12;
+            this.kchkUseProgressValueAsText.Values.Text = "Use progress value as text";
             // 
             // kryptonProgressBar1
             // 
-            kryptonProgressBar1.Location = new System.Drawing.Point(269, 487);
-            kryptonProgressBar1.Margin = new System.Windows.Forms.Padding(4);
-            kryptonProgressBar1.Name = "kryptonProgressBar1";
-            kryptonProgressBar1.Size = new System.Drawing.Size(517, 32);
-            kryptonProgressBar1.StateCommon.Back.Color1 = System.Drawing.Color.Green;
-            kryptonProgressBar1.StateDisabled.Back.ColorStyle = Krypton.Toolkit.PaletteColorStyle.OneNote;
-            kryptonProgressBar1.StateNormal.Back.ColorStyle = Krypton.Toolkit.PaletteColorStyle.OneNote;
-            kryptonProgressBar1.TabIndex = 11;
-            kryptonProgressBar1.Text = "kryptonProgressBar1";
-            kryptonProgressBar1.Values.Text = "kryptonProgressBar1";
+            this.kryptonProgressBar1.Location = new System.Drawing.Point(202, 396);
+            this.kryptonProgressBar1.Name = "kryptonProgressBar1";
+            this.kryptonProgressBar1.Size = new System.Drawing.Size(388, 26);
+            this.kryptonProgressBar1.StateCommon.Back.Color1 = System.Drawing.Color.Green;
+            this.kryptonProgressBar1.StateDisabled.Back.ColorStyle = Krypton.Toolkit.PaletteColorStyle.OneNote;
+            this.kryptonProgressBar1.StateNormal.Back.ColorStyle = Krypton.Toolkit.PaletteColorStyle.OneNote;
+            this.kryptonProgressBar1.TabIndex = 11;
+            this.kryptonProgressBar1.Text = "kryptonProgressBar1";
+            this.kryptonProgressBar1.Values.Text = "kryptonProgressBar1";
             // 
             // ktrkProgressValues
             // 
-            ktrkProgressValues.Location = new System.Drawing.Point(269, 446);
-            ktrkProgressValues.Margin = new System.Windows.Forms.Padding(4);
-            ktrkProgressValues.Maximum = 100;
-            ktrkProgressValues.Name = "ktrkProgressValues";
-            ktrkProgressValues.Size = new System.Drawing.Size(517, 33);
-            ktrkProgressValues.TabIndex = 10;
-            ktrkProgressValues.TickStyle = System.Windows.Forms.TickStyle.Both;
-            ktrkProgressValues.ValueChanged += ktrkProgressValues_ValueChanged;
+            this.ktrkProgressValues.Location = new System.Drawing.Point(202, 362);
+            this.ktrkProgressValues.Maximum = 100;
+            this.ktrkProgressValues.Name = "ktrkProgressValues";
+            this.ktrkProgressValues.Size = new System.Drawing.Size(388, 33);
+            this.ktrkProgressValues.TabIndex = 10;
+            this.ktrkProgressValues.TickStyle = System.Windows.Forms.TickStyle.Both;
             // 
             // kryptonButton3
             // 
-            kryptonButton3.Location = new System.Drawing.Point(18, 357);
-            kryptonButton3.Margin = new System.Windows.Forms.Padding(4);
-            kryptonButton3.Name = "kryptonButton3";
-            kryptonButton3.Size = new System.Drawing.Size(245, 31);
-            kryptonButton3.TabIndex = 9;
-            kryptonButton3.Values.Text = "ThemeBrowser Form";
-            kryptonButton3.Click += kryptonButton3_Click;
+            this.kryptonButton3.Location = new System.Drawing.Point(14, 290);
+            this.kryptonButton3.Name = "kryptonButton3";
+            this.kryptonButton3.Size = new System.Drawing.Size(184, 25);
+            this.kryptonButton3.TabIndex = 9;
+            this.kryptonButton3.Values.Text = "ThemeBrowser Form";
             // 
             // kbtnIntegratedToolbar
             // 
-            kbtnIntegratedToolbar.Location = new System.Drawing.Point(19, 318);
-            kbtnIntegratedToolbar.Margin = new System.Windows.Forms.Padding(4);
-            kbtnIntegratedToolbar.Name = "kbtnIntegratedToolbar";
-            kbtnIntegratedToolbar.Size = new System.Drawing.Size(245, 31);
-            kbtnIntegratedToolbar.TabIndex = 8;
-            kbtnIntegratedToolbar.Values.Text = "Integrated Toolbar (Form5)";
-            kbtnIntegratedToolbar.Click += kbtnIntegratedToolbar_Click;
+            this.kbtnIntegratedToolbar.Location = new System.Drawing.Point(14, 258);
+            this.kbtnIntegratedToolbar.Name = "kbtnIntegratedToolbar";
+            this.kbtnIntegratedToolbar.Size = new System.Drawing.Size(184, 25);
+            this.kbtnIntegratedToolbar.TabIndex = 8;
+            this.kbtnIntegratedToolbar.Values.Text = "Integrated Toolbar (Form5)";
             // 
             // kbtnTestMessagebox
             // 
-            kbtnTestMessagebox.Location = new System.Drawing.Point(17, 279);
-            kbtnTestMessagebox.Margin = new System.Windows.Forms.Padding(4);
-            kbtnTestMessagebox.Name = "kbtnTestMessagebox";
-            kbtnTestMessagebox.Size = new System.Drawing.Size(247, 31);
-            kbtnTestMessagebox.TabIndex = 7;
-            kbtnTestMessagebox.Values.Text = "Test Messagebox";
-            kbtnTestMessagebox.Click += kbtnTestMessagebox_Click;
+            this.kbtnTestMessagebox.Location = new System.Drawing.Point(13, 227);
+            this.kbtnTestMessagebox.Name = "kbtnTestMessagebox";
+            this.kbtnTestMessagebox.Size = new System.Drawing.Size(185, 25);
+            this.kbtnTestMessagebox.TabIndex = 7;
+            this.kbtnTestMessagebox.Values.Text = "Test Messagebox";
             // 
             // kryptonButton2
             // 
-            kryptonButton2.Location = new System.Drawing.Point(17, 201);
-            kryptonButton2.Margin = new System.Windows.Forms.Padding(4);
-            kryptonButton2.Name = "kryptonButton2";
-            kryptonButton2.Size = new System.Drawing.Size(247, 31);
-            kryptonButton2.TabIndex = 6;
-            kryptonButton2.Values.Text = "Ribbon (Form3)";
-            kryptonButton2.Click += kryptonButton2_Click;
+            this.kryptonButton2.Location = new System.Drawing.Point(13, 163);
+            this.kryptonButton2.Name = "kryptonButton2";
+            this.kryptonButton2.Size = new System.Drawing.Size(185, 25);
+            this.kryptonButton2.TabIndex = 6;
+            this.kryptonButton2.Values.Text = "Ribbon (Form3)";
             // 
             // kryptonThemeComboBox1
             // 
-            kryptonThemeComboBox1.DisplayMember = "Key";
-            kryptonThemeComboBox1.DropDownWidth = 121;
-            kryptonThemeComboBox1.IntegralHeight = false;
-            kryptonThemeComboBox1.Location = new System.Drawing.Point(16, 15);
-            kryptonThemeComboBox1.Margin = new System.Windows.Forms.Padding(4);
-            kryptonThemeComboBox1.Name = "kryptonThemeComboBox1";
-            kryptonThemeComboBox1.Size = new System.Drawing.Size(247, 25);
-            kryptonThemeComboBox1.StateCommon.ComboBox.Border.DrawBorders = Krypton.Toolkit.PaletteDrawBorders.Top | Krypton.Toolkit.PaletteDrawBorders.Bottom | Krypton.Toolkit.PaletteDrawBorders.Left | Krypton.Toolkit.PaletteDrawBorders.Right;
-            kryptonThemeComboBox1.StateCommon.ComboBox.Content.TextH = Krypton.Toolkit.PaletteRelativeAlign.Near;
-            kryptonThemeComboBox1.TabIndex = 4;
-            kryptonThemeComboBox1.ValueMember = "Value";
+            this.kryptonThemeComboBox1.DisplayMember = "Key";
+            this.kryptonThemeComboBox1.DropDownWidth = 121;
+            this.kryptonThemeComboBox1.IntegralHeight = false;
+            this.kryptonThemeComboBox1.Location = new System.Drawing.Point(12, 12);
+            this.kryptonThemeComboBox1.Name = "kryptonThemeComboBox1";
+            this.kryptonThemeComboBox1.Size = new System.Drawing.Size(185, 21);
+            this.kryptonThemeComboBox1.StateCommon.ComboBox.Border.DrawBorders = ((Krypton.Toolkit.PaletteDrawBorders)((((Krypton.Toolkit.PaletteDrawBorders.Top | Krypton.Toolkit.PaletteDrawBorders.Bottom)
+            | Krypton.Toolkit.PaletteDrawBorders.Left)
+            | Krypton.Toolkit.PaletteDrawBorders.Right)));
+            this.kryptonThemeComboBox1.StateCommon.ComboBox.Content.TextH = Krypton.Toolkit.PaletteRelativeAlign.Near;
+            this.kryptonThemeComboBox1.TabIndex = 4;
+            this.kryptonThemeComboBox1.ValueMember = "Value";
             // 
             // kryptonButton1
             // 
-            kryptonButton1.KryptonContextMenu = kryptonContextMenu1;
-            kryptonButton1.Location = new System.Drawing.Point(16, 162);
-            kryptonButton1.Margin = new System.Windows.Forms.Padding(4);
-            kryptonButton1.Name = "kryptonButton1";
-            kryptonButton1.ShowSplitOption = true;
-            kryptonButton1.Size = new System.Drawing.Size(247, 31);
-            kryptonButton1.TabIndex = 5;
-            kryptonButton1.Values.Text = "Button (form2)";
-            kryptonButton1.Click += kryptonButton1_Click;
+            this.kryptonButton1.KryptonContextMenu = this.kryptonContextMenu1;
+            this.kryptonButton1.Location = new System.Drawing.Point(12, 132);
+            this.kryptonButton1.Name = "kryptonButton1";
+            this.kryptonButton1.ShowSplitOption = true;
+            this.kryptonButton1.Size = new System.Drawing.Size(185, 25);
+            this.kryptonButton1.TabIndex = 5;
+            this.kryptonButton1.Values.Text = "Button (form2)";
             // 
             // kryptonContextMenu1
             // 
-            kryptonContextMenu1.Items.AddRange(new Krypton.Toolkit.KryptonContextMenuItemBase[] { kryptonContextMenuItems1 });
+            this.kryptonContextMenu1.Items.AddRange(new Krypton.Toolkit.KryptonContextMenuItemBase[] {
+            this.kryptonContextMenuItems1});
             // 
             // kryptonContextMenuItems1
             // 
-            kryptonContextMenuItems1.Items.AddRange(new Krypton.Toolkit.KryptonContextMenuItemBase[] { kryptonContextMenuItem1, kryptonContextMenuItem2, kryptonContextMenuItem3 });
+            this.kryptonContextMenuItems1.Items.AddRange(new Krypton.Toolkit.KryptonContextMenuItemBase[] {
+            this.kryptonContextMenuItem1,
+            this.kryptonContextMenuItem2,
+            this.kryptonContextMenuItem3});
             // 
             // kryptonContextMenuItem1
             // 
-            kryptonContextMenuItem1.Text = "Menu Item 1";
-            kryptonContextMenuItem1.ToolTipValues.Description = "Multi lin Description:\r\n- Match thing 1\r\n- Match thing 2\r\n\r\nPlacement mode: Right";
-            kryptonContextMenuItem1.ToolTipValues.EnableToolTips = true;
-            kryptonContextMenuItem1.ToolTipValues.Heading = "Menu Tooltip Heading";
-            kryptonContextMenuItem1.ToolTipValues.ToolTipPosition.PlacementMode = Krypton.Toolkit.PlacementMode.Right;
+            this.kryptonContextMenuItem1.Text = "Menu Item 1";
+            this.kryptonContextMenuItem1.ToolTipValues.Description = "Multi lin Description:\r\n- Match thing 1\r\n- Match thing 2\r\n\r\nPlacement mode: Right" +
+    "";
+            this.kryptonContextMenuItem1.ToolTipValues.EnableToolTips = true;
+            this.kryptonContextMenuItem1.ToolTipValues.Heading = "Menu Tooltip Heading";
+            this.kryptonContextMenuItem1.ToolTipValues.ToolTipPosition.PlacementMode = Krypton.Toolkit.PlacementMode.Right;
             // 
             // kryptonContextMenuItem2
             // 
-            kryptonContextMenuItem2.Text = "Menu Item 2";
+            this.kryptonContextMenuItem2.Text = "Menu Item 2";
             // 
             // kryptonContextMenuItem3
             // 
-            kryptonContextMenuItem3.Text = "Menu Item 3";
+            this.kryptonContextMenuItem3.Text = "Menu Item 3";
             // 
             // kryptonListBox1
             // 
-            kryptonListBox1.Location = new System.Drawing.Point(271, 11);
-            kryptonListBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
-            kryptonListBox1.Name = "kryptonListBox1";
-            kryptonListBox1.Size = new System.Drawing.Size(517, 426);
-            kryptonListBox1.TabIndex = 2;
+            this.kryptonListBox1.Location = new System.Drawing.Point(203, 9);
+            this.kryptonListBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.kryptonListBox1.Name = "kryptonListBox1";
+            this.kryptonListBox1.Size = new System.Drawing.Size(388, 346);
+            this.kryptonListBox1.TabIndex = 2;
             // 
             // kryptonTextBox1
             // 
-            kryptonTextBox1.Location = new System.Drawing.Point(77, 128);
-            kryptonTextBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
-            kryptonTextBox1.Name = "kryptonTextBox1";
-            kryptonTextBox1.Size = new System.Drawing.Size(100, 27);
-            kryptonTextBox1.TabIndex = 1;
-            kryptonTextBox1.Text = "kryptonTextBox1";
-            kryptonTextBox1.Click += kryptonTextBox1_Click;
-            kryptonTextBox1.DoubleClick += kryptonTextBox1_DoubleClick;
-            kryptonTextBox1.KeyDown += kryptonTextBox1_KeyDown;
-            kryptonTextBox1.KeyPress += kryptonTextBox1_KeyPress;
-            kryptonTextBox1.KeyUp += kryptonTextBox1_KeyUp;
-            kryptonTextBox1.MouseClick += kryptonTextBox1_MouseClick;
-            kryptonTextBox1.MouseDoubleClick += kryptonTextBox1_MouseDoubleClick;
-            kryptonTextBox1.PreviewKeyDown += kryptonTextBox1_PreviewKeyDown;
-            kryptonTextBox1.Validating += kryptonTextBox1_Validating;
-            kryptonTextBox1.Validated += kryptonTextBox1_Validated;
+            this.kryptonTextBox1.Location = new System.Drawing.Point(58, 104);
+            this.kryptonTextBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.kryptonTextBox1.Name = "kryptonTextBox1";
+            this.kryptonTextBox1.Size = new System.Drawing.Size(75, 23);
+            this.kryptonTextBox1.TabIndex = 1;
+            this.kryptonTextBox1.Text = "kryptonTextBox1";
             // 
             // textBox1
             // 
-            textBox1.Location = new System.Drawing.Point(73, 60);
-            textBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
-            textBox1.Name = "textBox1";
-            textBox1.Size = new System.Drawing.Size(100, 22);
-            textBox1.TabIndex = 0;
-            textBox1.Click += textBox1_Click;
-            textBox1.MouseClick += textBox1_MouseClick;
-            textBox1.DoubleClick += textBox1_DoubleClick;
-            textBox1.KeyDown += textBox1_KeyDown;
-            textBox1.KeyPress += textBox1_KeyPress;
-            textBox1.KeyUp += textBox1_KeyUp;
-            textBox1.MouseDoubleClick += textBox1_MouseDoubleClick;
-            textBox1.PreviewKeyDown += textBox1_PreviewKeyDown;
-            textBox1.Validating += textBox1_Validating;
-            textBox1.Validated += textBox1_Validated;
+            this.textBox1.Location = new System.Drawing.Point(55, 49);
+            this.textBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.textBox1.Name = "textBox1";
+            this.textBox1.Size = new System.Drawing.Size(76, 20);
+            this.textBox1.TabIndex = 0;
             // 
             // kryptonCustomPaletteBase1
             // 
-            kryptonCustomPaletteBase1.BaseFont = new System.Drawing.Font("Segoe UI", 9F);
-            kryptonCustomPaletteBase1.BaseFontSize = 9F;
-            kryptonCustomPaletteBase1.BasePaletteType = Krypton.Toolkit.BasePaletteType.Custom;
-            kryptonCustomPaletteBase1.ThemeName = null;
-            kryptonCustomPaletteBase1.UseKryptonFileDialogs = true;
+            this.kryptonCustomPaletteBase1.BaseFont = new System.Drawing.Font("Segoe UI", 9F);
+            this.kryptonCustomPaletteBase1.BaseFontSize = 9F;
+            this.kryptonCustomPaletteBase1.BasePaletteType = Krypton.Toolkit.BasePaletteType.Custom;
+            this.kryptonCustomPaletteBase1.ThemeName = null;
+            this.kryptonCustomPaletteBase1.UseKryptonFileDialogs = true;
             // 
             // kcmdMessageboxTest
             // 
-            kcmdMessageboxTest.Text = "kryptonCommand1";
-            kcmdMessageboxTest.Execute += kcmdMessageboxTest_Execute;
+            this.kcmdMessageboxTest.Text = "kryptonCommand1";
             // 
             // buttonSpecAny1
             // 
-            buttonSpecAny1.UniqueName = "dd7cce66eb3d4d97a9536f5d03a58dd5";
+            this.buttonSpecAny1.UniqueName = "dd7cce66eb3d4d97a9536f5d03a58dd5";
             // 
             // buttonSpecAny2
             // 
-            buttonSpecAny2.UniqueName = "01c062f324854832b96667cacd7d1a77";
+            this.buttonSpecAny2.UniqueName = "01c062f324854832b96667cacd7d1a77";
             // 
             // buttonSpecAny3
             // 
-            buttonSpecAny3.UniqueName = "be5331ab7c484486a10d7094d39cd05e";
+            this.buttonSpecAny3.UniqueName = "be5331ab7c484486a10d7094d39cd05e";
             // 
             // buttonSpecAny4
             // 
-            buttonSpecAny4.UniqueName = "e8dbb053058c44ed9c8f2b7c51ec1261";
+            this.buttonSpecAny4.UniqueName = "e8dbb053058c44ed9c8f2b7c51ec1261";
             // 
             // buttonSpecAny5
             // 
-            buttonSpecAny5.UniqueName = "abaa2bdd24704458afca577b2da97ed2";
+            this.buttonSpecAny5.UniqueName = "abaa2bdd24704458afca577b2da97ed2";
             // 
             // buttonSpecAny6
             // 
-            buttonSpecAny6.UniqueName = "17c90049490f478399fdad972722b324";
+            this.buttonSpecAny6.UniqueName = "17c90049490f478399fdad972722b324";
             // 
             // buttonSpecAny7
             // 
-            buttonSpecAny7.UniqueName = "8ec878c93ad347528009c3a5e4a71345";
+            this.buttonSpecAny7.UniqueName = "8ec878c93ad347528009c3a5e4a71345";
             // 
             // buttonSpecAny8
             // 
-            buttonSpecAny8.UniqueName = "1cd972225f3441b5ade70ac3c02d3949";
+            this.buttonSpecAny8.UniqueName = "1cd972225f3441b5ade70ac3c02d3949";
             // 
             // buttonSpecAny9
             // 
-            buttonSpecAny9.UniqueName = "1d2714d06d564ef580dff961d56d525a";
+            this.buttonSpecAny9.UniqueName = "1d2714d06d564ef580dff961d56d525a";
             // 
             // buttonSpecAny10
             // 
-            buttonSpecAny10.UniqueName = "8b3bf64a4cd7424d9479e809c7485ae7";
+            this.buttonSpecAny10.UniqueName = "8b3bf64a4cd7424d9479e809c7485ae7";
             // 
             // kryptonIntegratedToolbarPrintCommand1
             // 
-            kryptonIntegratedToolbarPrintCommand1.Text = "Print";
+            this.kryptonIntegratedToolbarPrintCommand1.Text = "Print";
+            // 
+            // kryptonButton5
+            // 
+            this.kryptonButton5.Location = new System.Drawing.Point(14, 353);
+            this.kryptonButton5.Name = "kryptonButton5";
+            this.kryptonButton5.Size = new System.Drawing.Size(183, 25);
+            this.kryptonButton5.TabIndex = 1;
+            this.kryptonButton5.Values.Text = "Powered by Control";
+            this.kryptonButton5.Click += new System.EventHandler(this.kryptonButton5_Click);
             // 
             // Form1
             // 
-            AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
-            AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            ButtonSpecs.Add(buttonSpecAny1);
-            ButtonSpecs.Add(buttonSpecAny2);
-            ButtonSpecs.Add(buttonSpecAny3);
-            ButtonSpecs.Add(buttonSpecAny4);
-            ButtonSpecs.Add(buttonSpecAny5);
-            ButtonSpecs.Add(buttonSpecAny6);
-            ButtonSpecs.Add(buttonSpecAny7);
-            ButtonSpecs.Add(buttonSpecAny8);
-            ButtonSpecs.Add(buttonSpecAny9);
-            ButtonSpecs.Add(buttonSpecAny10);
-            ClientSize = new System.Drawing.Size(800, 693);
-            Controls.Add(kryptonPanel1);
-            FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
-            Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
-            Name = "Form1";
-            Text = "Form1";
-            WindowState = System.Windows.Forms.FormWindowState.Maximized;
-            ((System.ComponentModel.ISupportInitialize)kryptonPanel1).EndInit();
-            kryptonPanel1.ResumeLayout(false);
-            kryptonPanel1.PerformLayout();
-            ((System.ComponentModel.ISupportInitialize)kryptonThemeComboBox1).EndInit();
-            ResumeLayout(false);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ButtonSpecs.Add(this.buttonSpecAny1);
+            this.ButtonSpecs.Add(this.buttonSpecAny2);
+            this.ButtonSpecs.Add(this.buttonSpecAny3);
+            this.ButtonSpecs.Add(this.buttonSpecAny4);
+            this.ButtonSpecs.Add(this.buttonSpecAny5);
+            this.ButtonSpecs.Add(this.buttonSpecAny6);
+            this.ButtonSpecs.Add(this.buttonSpecAny7);
+            this.ButtonSpecs.Add(this.buttonSpecAny8);
+            this.ButtonSpecs.Add(this.buttonSpecAny9);
+            this.ButtonSpecs.Add(this.buttonSpecAny10);
+            this.ClientSize = new System.Drawing.Size(600, 563);
+            this.Controls.Add(this.kryptonPanel1);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            this.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.Name = "Form1";
+            this.Text = "Form1";
+            this.WindowState = System.Windows.Forms.FormWindowState.Maximized;
+            this.Load += new System.EventHandler(this.Form1_Load);
+            ((System.ComponentModel.ISupportInitialize)(this.kryptonPanel1)).EndInit();
+            this.kryptonPanel1.ResumeLayout(false);
+            this.kryptonPanel1.PerformLayout();
+            ((System.ComponentModel.ISupportInitialize)(this.kryptonThemeComboBox1)).EndInit();
+            this.ResumeLayout(false);
+
         }
 
         #endregion
@@ -403,5 +382,6 @@
         private Krypton.Toolkit.KryptonCheckBox kchkUseProgressValueAsText;
         private Krypton.Toolkit.KryptonButton kbtnVisualStudio2010Theme;
         private Krypton.Toolkit.KryptonButton kryptonButton4;
+        private Krypton.Toolkit.KryptonButton kryptonButton5;
     }
 }